// architecture/architecture-portal/src/app/resources/metrics/types/MetricsApiCount.ts
export type MetricsApiList = {
    id: string
    name: string
}

export interface MetricsApiCount {
    api_published: number
    proposed_apis: number
    darb_approved_apis: number
    earb_approved_apis: number
    onboarded_catalog_apis: number
    design_certified_apis: number
    prod_certified_apis: number
    total_apis: number
    domain: string
    earb_approved_names: MetricsApiList[]
    darb_approved_names: MetricsApiList[]
    onboarded_catalog_names_api: MetricsApiList[]
    design_certified_names: MetricsApiList[]
    prod_certified_names: MetricsApiList[]
    prim_company_domain_id: string
    published: number
    typea_published: number
    typeb_published: number
    draft: number
    typea_draft: number
    typeb_draft: number
    proposed: number
    typea_proposed: number
    typeb_proposed: number
    onboarded: number
    typea_onboarded: number
    typeb_onboarded: number
    earb: number
    typea_earb: number
    typeb_earb: number
    typea_dARB_Approved: number
    darb: number
    typeb_darb: number
    typea_darb: number
    design: number
    typea_design: number
    typeb_design: number
    prod: number
    typea_prod: number
    typeb_prod: number
    proposedTypeACumSum?: number
    darbTypeACumSum?: number
    earbTypeACumSum: number
    designTypeACumSum?: number
    darbTypeBCumSum?: number
    earbTypeBCumSum?: number
    designTypeBCumSum?: number
    opsProposed?: number
    opsDarb?: number
    opsEarb?: number
    opsDesign?: number
    opsProd?: number
    opsProposedTypeA?: number
    opsDarbTypeA?: number
    opsEarbTypeA?: number
    opsDesignTypeA?: number
    opsProdTypeA?: number
    opsProposedTypeB?: number
    opsDarbTypeB?: number
    opsEarbTypeB?: number
    opsDesignTypeB?: number
    opsProdTypeB?: number
    percentage: number
    ecmi?: string
    ownersvp?: string
    unitcio?: string
    techowner?: string
    company_domain_ids?: string[]
    avg_design_quality_score?: number
    avg_operation_design_quality_score?: number
    median_design_quality_score?: number
    median_operation_design_quality_score?: number
}